pyproject.toml
[tool.poetry]
name = "aws-generated-data"
version = "0.1.0"
description = ""
authors = ["Christian Assing <cassing@redhat.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "aws_generated_data" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
typer = "^0.9.0"
rich = "^13.5.2"
pyyaml = "^6.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
flake8 = "^6.1.0"
pylint = "^2.17.5"
black = "^23.9.1"
pytest = "^7.4.2"
types-pyyaml = "^6.0.12.11"
types-requests = "^2.31.0.2"
types-beautifulsoup4 = "^4.12.0.6"
requests-mock = "^1.11.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
agd = 'aws_generated_data.__main__:app'

[tool.black]
line-length = 88
target-version = ['py311']
include = '^.*\.py$'

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 2
float_to_top = true

[tool.mypy]
plugins = "pydantic.mypy"
enable_error_code = "truthy-bool, redundant-expr"
no_implicit_optional = true
check_untyped_defs = true
warn_unused_ignores = true
show_error_codes = true
disallow_untyped_defs = true
disallow_incomplete_defs = true


[[tool.mypy.overrides]]
module = ["pyquery.*"]
ignore_missing_imports = true
